+++ docs/app/components/content/examples/table/TableExample.vue
@@ -190,6 +190,9 @@ const columns: TableColumn<Payment>[] = [
   {
     id: 'actions',
     meta: {
+      class: {
+        td: 'text-right'
+      },
       style: {
         td: {
           width: '20px',
@@ -291,20 +294,27 @@ const columns: TableColumn<Payment>[] = [
         })
       })
     },
-    cell: ({ row }) => h('div', { class: 'lowercase' }, row.getValue('email'))
+    meta: {
+      class: {
+        td: 'lowercase'
+      }
+    }
   },
   {
     accessorKey: 'amount',
-    header: () => h('div', { class: 'text-right' }, 'Amount'),
+    header: 'Amount',
+    meta: {
+      class: {
+        th: 'text-right',
+        td: 'text-right font-medium'
+      }
+    },
     cell: ({ row }) => {
       const amount = Number.parseFloat(row.getValue('amount'))
-
-      const formatted = new Intl.NumberFormat('en-US', {
+      return new Intl.NumberFormat('en-US', {
         style: 'currency',
         currency: 'EUR'
       }).format(amount)
-
-      return h('div', { class: 'text-right font-(--ui-font-weight-medium)' }, formatted)
     }
   }
 ]